Bumped API to ALPHA10, ALPHA11
Removed old API supports.
Bumped Plugin Version to v1.2.6
Changed Author due to the plugin developer’s inactivitey.
The main developer of this plugin is counted as inactive. He has not worked nor updated this plugin for 2 months, which in this case, he has been counted as being inactive. So we took over the project to support the latest API’s.